requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn[standard]
websockets